libmakepkg: Allow Git system config
On build.archlinux.org, we use a system-wide `/etc/gitconfig` to set

```gitconfig
[safe]
	directory = *
```

This is required for multiple users to share the same `SRCDEST`;
otherwise Git will refuse to operate on repos owned by other users.

Keep the user's global config suppressed but allow system-wide config.
We will assume this is less of a wild west of options that could
unintentionally affect git-archive output.