libalpm: Ignore database locks if the process that locked the database no longer exists
It's quite common to get the following message after aborting a pacman command, often after system restarts:

:: Synchronizing package databases...
error: failed to synchronize all databases (unable to lock database)

Right now, the locking mechanism is relatively rudimentary: if db.lck exists, then the database is considered locked. If something happens to the pacman process and it cannot remove the db.lck, it will be locked indefinitely, or until somebody Googles "pacman unable to lock database" and removes the lock.

This commit makes db.lck keep track of the process that locked it. If subsequent pacman processes cannot find the process that originally locked the database, the database will be considered unlocked.